+++ Editor/Interfaces/IAnimationModuleAnalyzer.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using Unity.ProjectAuditor.Editor.Core;
+using UnityEditor;
+
+namespace Unity.ProjectAuditor.Editor.Interfaces
+{
+    internal interface IAnimationModuleAnalyzer : IModuleAnalyzer
+    {
+    }
+}