+++ Editor/Modules/AnimationAnalyzer.cs
@@ -1,0 +1,19 @@
+using System.Collections.Generic;
+using System.IO;
+using Unity.ProjectAuditor.Editor.Core;
+using Unity.ProjectAuditor.Editor.Diagnostic;
+using Unity.ProjectAuditor.Editor.Interfaces;
+using UnityEditor;
+using UnityEngine;
+using UnityEngine.Profiling;
+using UnityEngine.Rendering;
+
+namespace Unity.ProjectAuditor.Editor.Modules
+{
+    class AnimationAnalyzer : IAnimationModuleAnalyzer
+    {
+        public void Initialize(ProjectAuditorModule module)
+        {
+        }
+    }
+}