+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * View-specific descriptions
 * AudioClip diagnostics
 * Material list view in shader module
+* AnimatorController reporting
+* AnimationClip reporting
+* Avatar reporting
+* AvatarMask reporting
 
 ### Changed
 * Improved issue description and suggestion strings, to make them clearer and more consistent